+++ inst/include/interface/rcpp/rcpp_objects/rcpp_data.hpp
@@ -184,10 +184,15 @@ class AgeCompDataInterface : public DataInterfaceBase {
    * @return A boolean of true.
    */
   virtual bool add_to_fims_tmb() {
+    #ifdef TMBAD_FRAMEWORK
+        this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
+        this->add_to_fims_tmb_internal<TMBAD_FIMS_TYPE>();
+    #else
     this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
     this->add_to_fims_tmb_internal<TMB_FIMS_FIRST_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_SECOND_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_THIRD_ORDER>();
+    #endif
 
     return true;
   }
@@ -295,10 +300,16 @@ class LengthCompDataInterface : public DataInterfaceBase {
    * @return A boolean of true.
    */
   virtual bool add_to_fims_tmb() {
+    #ifdef TMBAD_FRAMEWORK
+        this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
+        this->add_to_fims_tmb_internal<TMBAD_FIMS_TYPE>();
+    #else
     this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
     this->add_to_fims_tmb_internal<TMB_FIMS_FIRST_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_SECOND_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_THIRD_ORDER>();
+    #endif
+
     return true;
   }
 #endif
@@ -398,10 +409,15 @@ class IndexDataInterface : public DataInterfaceBase {
    * @return A boolean of true.
    */
   virtual bool add_to_fims_tmb() {
+    #ifdef TMBAD_FRAMEWORK
+        this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
+        this->add_to_fims_tmb_internal<TMBAD_FIMS_TYPE>();
+    #else
     this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
     this->add_to_fims_tmb_internal<TMB_FIMS_FIRST_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_SECOND_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_THIRD_ORDER>();
+    #endif
 
     return true;
   }
@@ -503,10 +519,15 @@ class LandingsDataInterface : public DataInterfaceBase {
    * @return A boolean of true.
    */
   virtual bool add_to_fims_tmb() {
+    #ifdef TMBAD_FRAMEWORK
+        this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
+        this->add_to_fims_tmb_internal<TMBAD_FIMS_TYPE>();
+    #else
     this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
     this->add_to_fims_tmb_internal<TMB_FIMS_FIRST_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_SECOND_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_THIRD_ORDER>();
+    #endif
 
     return true;
   }
