+++ inst/include/interface/rcpp/rcpp_objects/rcpp_fleet.hpp
@@ -874,12 +874,17 @@ public:
    * @return A boolean of true.
    */
   virtual bool add_to_fims_tmb() {
+    #ifdef TMBAD_FRAMEWORK
+        this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
+        this->add_to_fims_tmb_internal<TMBAD_FIMS_TYPE>();
+    #else
     FIMS_INFO_LOG("adding Fleet object to TMB");
     this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
     this->add_to_fims_tmb_internal<TMB_FIMS_FIRST_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_SECOND_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_THIRD_ORDER>();
-
+    #endif
+    
     return true;
   }
