+++ inst/include/interface/rcpp/rcpp_objects/rcpp_growth.hpp
@@ -216,10 +216,15 @@ class EWAAGrowthInterface : public GrowthInterfaceBase {
    * @return A boolean of true.
    */
   virtual bool add_to_fims_tmb() {
+        this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
+    #ifdef TMBAD_FRAMEWORK
+        this->add_to_fims_tmb_internal<TMBAD_FIMS_TYPE>();
+    #else
     this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
     this->add_to_fims_tmb_internal<TMB_FIMS_FIRST_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_SECOND_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_THIRD_ORDER>();
+    #endif
 
     return true;
   }
