+++ inst/include/interface/rcpp/rcpp_objects/rcpp_population.hpp
@@ -457,11 +457,15 @@ class PopulationInterface : public PopulationInterfaceBase {
    * @return A boolean of true.
    */
   virtual bool add_to_fims_tmb() {
-    FIMS_INFO_LOG("adding Population object to TMB");
+    #ifdef TMBAD_FRAMEWORK
+        this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
+        this->add_to_fims_tmb_internal<TMBAD_FIMS_TYPE>();
+    #else
     this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
     this->add_to_fims_tmb_internal<TMB_FIMS_FIRST_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_SECOND_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_THIRD_ORDER>();
+    #endif
 
     return true;
   }
