+++ inst/include/interface/rcpp/rcpp_objects/rcpp_selectivity.hpp
@@ -266,11 +266,15 @@ class LogisticSelectivityInterface : public SelectivityInterfaceBase {
    * @return A boolean of true.
    */
   virtual bool add_to_fims_tmb() {
-    FIMS_INFO_LOG("adding Selectivity object to TMB");
+    #ifdef TMBAD_FRAMEWORK
+        this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
+        this->add_to_fims_tmb_internal<TMBAD_FIMS_TYPE>();
+    #else
     this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
     this->add_to_fims_tmb_internal<TMB_FIMS_FIRST_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_SECOND_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_THIRD_ORDER>();
+    #endif
 
     return true;
   }
@@ -537,10 +541,15 @@ class DoubleLogisticSelectivityInterface : public SelectivityInterfaceBase {
    * @return A boolean of true.
    */
   virtual bool add_to_fims_tmb() {
+    #ifdef TMBAD_FRAMEWORK
+        this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
+        this->add_to_fims_tmb_internal<TMBAD_FIMS_TYPE>();
+    #else
     this->add_to_fims_tmb_internal<TMB_FIMS_REAL_TYPE>();
     this->add_to_fims_tmb_internal<TMB_FIMS_FIRST_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_SECOND_ORDER>();
     this->add_to_fims_tmb_internal<TMB_FIMS_THIRD_ORDER>();
+    #endif
 
     return true;
   }
