+++ src/FIMS.cpp
@@ -3,7 +3,8 @@
 
 #include "../inst/include/interface/rcpp/rcpp_interface.hpp"
 #include "../inst/include/interface/interface.hpp"
-#include "../inst/include/interface/init.hpp"
+#include "init.hpp"
+#include "fims_modules.hpp"
 #include "../inst/include/common/model.hpp"
 
 /// @cond
