+++ src/Makevars
@@ -1,4 +1,4 @@
 CXX_STD = CXX17
-PKG_CXXFLAGS = -DTMB_MODEL  -DTMB_EIGEN_DISABLE_WARNINGS
+PKG_CXXFLAGS = -DTMB_MODEL  -DTMB_EIGEN_DISABLE_WARNINGS  -DTMBAD_FRAMEWORK
 CXX17STD = -std=c++17 -w
 USE_CXX17 = "yes"
